pyproject.toml
[project]
name = "vector-ats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "markitdown[pdf]>=0.1.2",
    "pdfminer-six>=20250506",
    "pdfplumber>=0.11.7",
    "pymupdf>=1.26.3",
]
